+++ src/django_bird/components.py
@@ -8,14 +8,12 @@ from threading import Lock
 from typing import TYPE_CHECKING
 
 from cachetools import LRUCache
-from django.apps import apps
 from django.conf import settings
 from django.template.backends.django import Template as DjangoTemplate
 from django.template.base import Node
 from django.template.base import NodeList
 from django.template.base import TextNode
 from django.template.context import Context
-from django.template.engine import Engine
 from django.template.exceptions import TemplateDoesNotExist
 from django.template.loader import select_template
 
@@ -28,6 +26,7 @@ from django_bird.templatetags.tags.slot import SlotNode
 
 from .conf import app_settings
 from .staticfiles import AssetType
+from .templates import get_component_directories
 from .templates import get_template_names
 
 if TYPE_CHECKING:
@@ -78,15 +77,16 @@ class Component:
     def from_name(cls, name: str):
         template_names = get_template_names(name)
         template = select_template(template_names)
-
-        assets: set[Asset] = set()
-        template_path = Path(template.template.origin.name)
-        for asset_type in AssetType:
-            asset_path = template_path.with_suffix(asset_type.ext)
-            if asset_path.exists():
-                asset = Asset.from_path(asset_path)
-                assets.add(asset)
-
+        assets: list[Asset] = [
+            asset
+            for asset_type in AssetType
+            if (
+                asset := Asset.from_path(
+                    Path(template.template.origin.name), asset_type
+                )
+            )
+            is not None
+        ]
         return cls(name=name, template=template, assets=frozenset(assets))
 
 
@@ -182,32 +182,8 @@ class ComponentRegistry:
     def __init__(self, maxsize: int = 100):
         self._components: LRUCache[str, Component] = LRUCache(maxsize=maxsize)
 
-    def get_dirs(self):
-        engine = Engine.get_default()
-        dirs: list[str | Path] = list(engine.dirs)
-
-        for app_config in apps.get_app_configs():
-            template_dir = Path(app_config.path) / "templates"
-            if template_dir.is_dir():
-                dirs.append(template_dir)
-
-        base_dir = getattr(settings, "BASE_DIR", None)
-
-        if base_dir is not None:
-            root_template_dir = Path(base_dir) / "templates"
-            if root_template_dir.is_dir():
-                dirs.append(root_template_dir)
-
-        return dirs
-
     def discover_components(self) -> None:
-        template_dirs = self.get_dirs()
-
-        component_dirs = [
-            Path(template_dir) / component_dir
-            for template_dir in template_dirs
-            for component_dir in app_settings.COMPONENT_DIRS + ["bird"]
-        ]
+        component_dirs = get_component_directories()
 
         for component_dir in component_dirs:
             component_dir = Path(component_dir)
@@ -240,13 +216,13 @@ class ComponentRegistry:
         self._components[name] = component
         return component
 
-    def get_assets(self, asset_type: AssetType) -> frozenset[Asset]:
-        assets: set[Asset] = set()
-        for component in self._components.values():
-            for asset in component.assets:
-                if asset.type == asset_type:
-                    assets.add(asset)
-        return frozenset(assets)
+    def get_assets(self, asset_type: AssetType | None = None) -> frozenset[Asset]:
+        return frozenset(
+            asset
+            for component in self._components.values()
+            for asset in component.assets
+            if asset_type is None or asset.type == asset_type
+        )
 
 
 components = ComponentRegistry()